only filelogging

+++ lib/logging.rb
@@ -7,25 +7,31 @@ end
 
 module Logging
   
+  
   def log(msg)
     @log ||= Logger.new(STDOUT)
     @log.level ||= Logger::DEBUG          
     @log_file ||= 'dryml_template'
+    @overwrite = false
 #    @old_template ||= ''
     
     if @template_path # && @template_path.include?('app/views/') || !@template_path.include?('views/taglibs')
       file = @template_path + '.log'
-      log_f(msg, file)
+      if File.exist?(@template_path)
+        log_f(msg, file)
+      end
     else
-       @log.debug msg
+       # @log.debug msg
        log_f(msg)
     end
   end  
 
   def log_f(txt, file = nil)
     file = file || (RAILS_ROOT + "/log/#{@log_file}.log")
-    open(file, "a+") do |f|
-      f.puts txt
+    unless File.exist?(file) && @overwrite
+      open(file, "a+") do |f|
+        f.puts txt
+      end
     end
   end
     
@@ -35,8 +41,10 @@ module SaveErb
   def save_erb_file(erb, file = nil)
     file = file || (RAILS_ROOT + "/" + @template_path + ".erb")
     unless (File.basename(@template_path, '.dryml') == 'application') || @template_path.include?('taglibs')
-      File.open(file, "w") do |erbfile|
-        erbfile.syswrite(erb)
+      unless File.exist?(file) && @overwrite      
+        File.open(file, "w") do |erbfile|
+          erbfile.syswrite(erb)
+        end
       end
     end
     erb
